move collect_doc_ids_from_manifest
