+++ plugins/woocommerce-docs/src/Manifest/ManifestProcessor.php
@@ -59,5 +59,23 @@ class ManifestProcessor {
 			}
 		}
 	}
+
+	/**
+	 * Recusively collect post IDs from a manifest.
+	 *
+	 * @param Object $manifest The manifest to process.
+	 */
+	public static function collect_doc_ids_from_manifest( $manifest ) {
+		$doc_ids = array();
+		foreach ( $manifest['categories'] as $category ) {
+			foreach ( $category['posts'] as $post ) {
+				$doc_ids[] = $post['id'];
+			}
+			$subcategory_ids = self::collect_doc_ids_from_manifest( $category );
+			$doc_ids         = array_merge( $doc_ids, $subcategory_ids );
+		}
+
+		return $doc_ids;
+	}
 }
